# contrats/openapi/api_moteur.openapi.json
{"openapi":"3.1.0","info":{"title":"API Moteur - Conseil des ministres","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/version":{"get":{"summary":"Version","operationId":"version_version_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/parties":{"post":{"summary":"Creer Partie","operationId":"creer_partie_parties_post","parameters":[{"name":"x-correlation-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Correlation-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequetePartie"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReponseEtat"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/parties/{partie_id}/etat":{"get":{"summary":"Lire Etat","operationId":"lire_etat_parties__partie_id__etat_get","parameters":[{"name":"partie_id","in":"path","required":true,"schema":{"type":"string","title":"Partie Id"}},{"name":"x-correlation-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Correlation-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReponseEtat"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/parties/{partie_id}/joueurs":{"post":{"summary":"Ajouter Joueur","operationId":"ajouter_joueur_parties__partie_id__joueurs_post","parameters":[{"name":"partie_id","in":"path","required":true,"schema":{"type":"string","title":"Partie Id"}},{"name":"x-correlation-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Correlation-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequeteJoueur"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReponseEtat"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/parties/{partie_id}/actions":{"post":{"summary":"Soumettre Action","operationId":"soumettre_action_parties__partie_id__actions_post","parameters":[{"name":"partie_id","in":"path","required":true,"schema":{"type":"string","title":"Partie Id"}},{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}},{"name":"x-correlation-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Correlation-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequeteAction"}}}},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReponseEtat"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/parties/{partie_id}/terminer":{"post":{"summary":"Terminer Partie","operationId":"terminer_partie_parties__partie_id__terminer_post","parameters":[{"name":"partie_id","in":"path","required":true,"schema":{"type":"string","title":"Partie Id"}},{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}},{"name":"x-correlation-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Correlation-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequeteTerminerPartie"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReponseEtat"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ConfigurationPartie":{"properties":{"politique_timeout_partie":{"anyOf":[{"$ref":"#/components/schemas/PolitiqueTimeoutPartie"},{"type":"null"}]}},"type":"object","title":"ConfigurationPartie"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PolitiqueTimeoutPartie":{"properties":{"version":{"type":"integer","title":"Version","default":1},"active":{"type":"boolean","title":"Active","default":true},"delai_inactivite_secondes":{"type":"integer","minimum":1.0,"title":"Delai Inactivite Secondes","default":3600}},"type":"object","title":"PolitiqueTimeoutPartie"},"ReponseEtat":{"properties":{"partie_id":{"type":"string","title":"Partie Id"},"etat":{"additionalProperties":true,"type":"object","title":"Etat"}},"type":"object","required":["partie_id","etat"],"title":"ReponseEtat"},"RequeteAction":{"properties":{"acteur":{"type":"string","title":"Acteur"},"type_action":{"type":"string","title":"Type Action"},"donnees":{"additionalProperties":true,"type":"object","title":"Donnees"}},"type":"object","required":["acteur","type_action"],"title":"RequeteAction"},"RequeteJoueur":{"properties":{"joueur_id":{"type":"string","title":"Joueur Id"},"pseudo":{"type":"string","title":"Pseudo"}},"type":"object","required":["joueur_id","pseudo"],"title":"RequeteJoueur"},"RequetePartie":{"properties":{"partie_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Partie Id"},"nom":{"type":"string","title":"Nom"},"joueurs":{"additionalProperties":true,"type":"object","title":"Joueurs"},"seed":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Seed"},"skin_jeu":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Skin Jeu","default":"minimal"},"configuration_partie":{"$ref":"#/components/schemas/ConfigurationPartie"},"options":{"additionalProperties":true,"type":"object","title":"Options"}},"type":"object","required":["nom","joueurs"],"title":"RequetePartie"},"RequeteTerminerPartie":{"properties":{"raison":{"type":"string","title":"Raison","default":"FIN_INCONNUE"}},"type":"object","title":"RequeteTerminerPartie"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}